Update `Card` docs
