+++ polaris.shopify.com/content/components/deprecated/card.md
@@ -26,55 +26,55 @@ status:
     Card was built prior to layout primitives like box, inline, and alpha stack.
     The new layout primitives can be used in combination to achieve similar results to card, or the existing legacy card.
 examples:
-  - fileName: card-default.tsx
+  - fileName: deprecated-card-default.tsx
     title: Default
     description: Use when you have a simple message to communicate to merchants that doesn’t require any secondary steps.
-  - fileName: card-with-header-actions.tsx
+  - fileName: deprecated-card-with-header-actions.tsx
     title: With header actions
     description: Use for less important card actions, or actions merchants may do before reviewing the contents of the card. For example, merchants may want to add items to a card containing a long list, or enter a customer’s new address.
-  - fileName: card-with-footer-actions.tsx
+  - fileName: deprecated-card-with-footer-actions.tsx
     title: With footer actions
     description: Use footer actions for a card’s most important actions, or actions merchants should do after reviewing the contents of the card. For example, merchants should review the contents of a shipment before an important action like adding tracking information. Footer actions can be left or right aligned with the `footerActionAlignment` prop.
-  - fileName: card-with-multiple-footer-actions.tsx
+  - fileName: deprecated-card-with-multiple-footer-actions.tsx
     title: With multiple footer actions
     description: When multiple secondary footer actions are provided, they will render in an action list popover activated by a disclosure button. The disclosure button text can be customized with the `secondaryFooterActionsDisclosureText` prop.
-  - fileName: card-with-custom-footer-actions.tsx
+  - fileName: deprecated-card-with-custom-footer-actions.tsx
     title: With custom footer actions
     description: Use to present actionable content that is optional or not the primary purpose of the page.
-  - fileName: card-with-destructive-footer-action.tsx
+  - fileName: deprecated-card-with-destructive-footer-action.tsx
     title: With destructive footer action
     description: Use when a card action will delete merchant data or be otherwise difficult to recover from.
-  - fileName: card-with-multiple-sections.tsx
+  - fileName: deprecated-card-with-multiple-sections.tsx
     title: With multiple sections
     description: Use when you have two related but distinct pieces of information to communicate to merchants. Multiple sections can help break up complicated concepts to make them easier to scan and understand.
-  - fileName: card-with-multiple-titled-sections.tsx
+  - fileName: deprecated-card-with-multiple-titled-sections.tsx
     title: With multiple titled sections
     description: Use when you have two related but distinct pieces of information to communicate to merchants that are complex enough to require a title to introduce them.
-  - fileName: card-with-sections-and-actions.tsx
+  - fileName: deprecated-card-with-sections-and-actions.tsx
     title: With sections and actions
     description: Use when your card section has actions that apply only to that section.
-  - fileName: card-with-subsection.tsx
+  - fileName: deprecated-card-with-subsection.tsx
     title: With subsection
     description: Use when your card sections need further categorization.
-  - fileName: card-with-destructive-action.tsx
+  - fileName: deprecated-card-with-destructive-action.tsx
     title: With destructive action
     description: Use when a card action applies only to one section and will delete merchant data or be otherwise difficult to recover from.
-  - fileName: card-with-a-subdued-section.tsx
+  - fileName: deprecated-card-with-a-subdued-section.tsx
     title: With a subdued section
     description: Use to indicate when one of the sections in your card contains inactive or disabled content.
-  - fileName: card-with-subdued-for-secondary-content.tsx
+  - fileName: deprecated-card-with-subdued-for-secondary-content.tsx
     title: With subdued for secondary content
     description: Use for content that you want to deprioritize. Subdued cards don’t stand out as much as cards with white backgrounds so don’t use them for information or actions that are critical to merchants.
-  - fileName: card-with-separate-header.tsx
+  - fileName: deprecated-card-with-separate-header.tsx
     title: With separate header
     description: Use to be able to use custom React elements as header content.
-  - fileName: card-with-custom-react-node-title.tsx
+  - fileName: deprecated-card-with-custom-react-node-title.tsx
     title: With custom React Node title
     description: Use to render custom content such as icons, links, or buttons in a card section’s header.
-  - fileName: card-with-all-elements.tsx
+  - fileName: deprecated-card-with-all-elements.tsx
     title: With all elements
     description: Use as a broad example that includes most props available to card.
-  - fileName: card-with-flushed-sections.tsx
+  - fileName: deprecated-card-with-flushed-sections.tsx
     title: With flushed sections
     description: Use when you need further control over the spacing of your card sections.
 ---
