--- polaris.shopify.com/pages/examples/alpha-card-default.tsx
@@ -1,15 +1,0 @@
-import {AlphaCard, Text} from '@shopify/polaris';
-import React from 'react';
-import {withPolarisExample} from '../../src/components/PolarisExampleWrapper';
-
-function AlphaCardExample() {
-  return (
-    <AlphaCard>
-      <Text as="h2" variant="bodyMd">
-        Content inside a card
-      </Text>
-    </AlphaCard>
-  );
-}
-
-export default withPolarisExample(AlphaCardExample);
