+++ polaris.shopify.com/pages/examples/card-with-varying-padding.tsx
@@ -25,7 +25,7 @@ const Placeholder = ({label = '', height = 'auto', width = 'auto'}) => {
   return (
     <div
       style={{
-        background: '#7B47F1',
+        background: '#20828D',
         height: height,
         width: width,
       }}
