+++ polaris.shopify.com/pages/examples/deprecated-card-default.tsx
@@ -1,0 +1,13 @@
+import {Card} from '@shopify/polaris';
+import React from 'react';
+import {withPolarisExample} from '../../src/components/PolarisExampleWrapper';
+
+function CardExample() {
+  return (
+    <Card title="Online store dashboard" sectioned>
+      <p>View a summary of your online store’s performance.</p>
+    </Card>
+  );
+}
+
+export default withPolarisExample(CardExample);
