+++ polaris.shopify.com/pages/examples/deprecated-card-with-custom-footer-actions.tsx
