+++ polaris.shopify.com/pages/examples/deprecated-card-with-destructive-action.tsx
