+++ polaris.shopify.com/pages/examples/deprecated-card-with-destructive-footer-action.tsx
