+++ polaris.shopify.com/pages/examples/deprecated-card-with-sections-and-actions.tsx
